Fixat den allvarliga buggen med att kunna manipulera och ta bort andras

+++ Feed/DeletePost.php
@@ -1,18 +1,34 @@
 <?php
 
+session_start();
+
 require_once('Model/Dao/PostRepository.php');
+require_once('Model/LoginModel.php');
 
 $postRepository = new PostRepository();
+$loginModel = new LoginModel();
 
-if (isset($_POST['image_name']) && empty($_POST['image_name']) == false) 
-{
-	unlink("View/Images/" . $_POST['image_name']);
-}
+$userId = $loginModel->getId();
+$posts = $postRepository->GetUsersPosts($userId);
 
-if (isset($_POST["feed_id"]) && strlen($_POST['feed_id']) > 0 && is_numeric($_POST['feed_id']))
+foreach ($posts as $post) 
 {
-	$feed_id = filter_var($_POST['feed_id'], FILTER_SANITIZE_NUMBER_INT); 
-	
-	$postRepository->DeletePost($feed_id);
+	if ($post['id'] == $_POST["feed_id"]) 
+	{
+		if (isset($_POST['image_name']) && empty($_POST['image_name']) == false) 
+		{
+			unlink("View/Images/" . $_POST['image_name']);
+		}
+
+		if (isset($_POST["feed_id"]) && strlen($_POST['feed_id']) > 0 && is_numeric($_POST['feed_id']))
+		{
+			$feed_id = filter_var($_POST['feed_id'], FILTER_SANITIZE_NUMBER_INT); 
+			
+			$postRepository->DeletePost($feed_id);
+
+			// Lyckat så responsa med 1
+			echo 1;
+		}
+	}
 }
